model-deployment-lecture/Deploy ML1/backend/requirements.txt
waitress
pandas
flask
tensorflow-cpu
scikit-learn==1.0.2